Helm: Reworked Helm to support more than AWS IAM Roles
This changes how the helm chart handles some of the stuff, so there's
some changes in the docs to reflect that.

The IAM Role is still the preferred way to work with AWS but it's
possible that a cluster doesn't have the ability to use IAM roles. The
changes to the helm chart allows for someone to set some env secrets.

Docs updated to highlight the two ways.