DNS-01 solver for cert-manager
Implement the basic functionality for Phonebook to integrate with
Cert-Manager. This allows any users that have a domain configured with
Phonebook to leverages cert-manager and Let's Encrypt to create
**wildcard** SSL Certificates dynamically.

The DNS-01 solver should work out of the box for any supported Provider.

This means that any provider that gets added to Phonebook will
automatically support dynamic SSL certificate, including wildcard
certificates.