fix(dnsrecords): Add errors to the event recorder
Errors were properly displayed on the condition, but because of the way
conditions are displayed, it's not easy to spot when investigating
issues.

In addition to having the error attached to the condition, this change
also adds the error to the event recorder. This way, the user can more
easily see the errors and then dig in the condition to know what
happened.