fix(k8s): Better handling of conflicts
Those conflict errors happen all the time when operator don't use
SSA(Server-side apply). Konditionner is built to try to catch those
errors early and as such, they shouldn't cause an issue with Phonebook.

However, they are annoying and the way the error is currently displayed
might lead to believe something is wrong with the operator. Instead,
let's wrap the error in a log message for now, until I either find a way
to fix it for good, or move to SSA.