BREAKING CHANGE: Rework Status for multi-provider support
Changed the DSNRecord status to be able to support n-number of
providers by providing a multi-level map. This is a bit of a rough
approach but for now I think it's better to have it working properly. I
believe the DNSRecord status is still unstable and possibly the Provider
interface needs to be reworked so it's more straightforward for the
clients to handle DNSRecord and updating the status.