fix(versions): Improve version management on release
Refactored tagged images in the codebase to use a variable that can be
set by a build flag so that github action can use a repository variable
to set those values.

This should make the release of new version easier as the github action
will use the tag's name and generate a version. From that point, all the
places that needs a version will source their version number off that
same environment variable.