feat(actions): Complete rework of github action for releases
This now uses a github action I've worked on to make deployment easier
with changelogs, helm charts, and version releases.

Now, it should be possible to only push a new tag to the repo and a new
release will be cut, with a helm package being added to the release as
well as an update to the index.yaml located in the gh-pages branch.

A changelog will also be created and updated on the main branch.

This is experimental and as such, a few kinks might need to be ironed
out as it's being used more.