fix(build): Export version to the right ldflag path
Go silently ignore values passed by `ldflags -X` and resulted in release
not carrying version numbers for the hardcoded values in the binary.

Maybe this way of packaging version should be revisited as it had a
bunch of issues with deployment, but maybe it's just one of those things
that are hard to get right but once it is, it becomes stable.